Fix MissingBodyAnnotationTest
The test in MissingBodyAnnotationTest has a request with a POJO as a body and no content type.

When converting this to a Azure Request prior to this PR we were simply calling toString() on
the POJO, which resulted in an invalid body in the request.

This PR attempts to convert the body to a usable form instead.

Only the tests are affected