Reduce flakiness in Azure http-test tck test
I suspect the issue is that we were finding a random port, and then passing this to Jetty.
In the mean time, something else may use this port.
And the heavier the load on the system, the greater the time between finding and using the port.
This PR switches to simply pass 0 to Jetty to allow it to find a port at the time it starts.
This is the same as we do in AWS.